Update getSurfaceFromX11Window to use ULong for window param
Previously, the window parameter was of type Long and converted to ULong internally. This change updates the parameter type directly to ULong, simplifying the code and improving type safety.

+++ demo/common/src/commonMain/kotlin/ext.kt
@@ -140,13 +140,13 @@ fun getSurfaceAndroidView(
 }
 
 
-fun getSurfaceFromX11Window(instance: WGPUInstance, display: NativeAddress, window: Long): WGPUSurface? = memoryScope { scope ->
+fun getSurfaceFromX11Window(instance: WGPUInstance, display: NativeAddress, window: ULong): WGPUSurface? = memoryScope { scope ->
 
     val surfaceDescriptor = WGPUSurfaceDescriptor.allocate(scope).apply {
         nextInChain = WGPUSurfaceSourceXlibWindow.allocate(scope).apply {
             chain.sType = WGPUSType_SurfaceSourceXlibWindow
             this.display = display
-            this.window = window.toULong()
+            this.window = window
         }.handler
     }
